# backend/api/openapi.yaml
openapi: 3.0.0
info:
  title: OpenTutor API
  version: 1.0.0
  description: |
    The OpenTutor API, for use with OpenTutor the frontend to
    provide various OpenTutor features.
  contact:
    name: OpenTutor API Support
    email: nathanjodoin@my.unt.edu
paths:
...
  /auth/login:
    post:
      summary: Log in as an existing user
      operationId: userLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserLogin"
      responses:
        "200":
          description: Login successful
          content:
            application/json: {}
          headers:
            x-session-token:
              schema:
                type: string
              description: The JWT associated with this authenticated session
        "400":
          description: User login failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorModel"

  # User register
  /auth/register:
    post:
      summary: Sign up as a new user
      operationId: userRegister
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/UserSignup"
      responses:
        "200":
          description: Sign up success
          content:
            application/json: {}
          headers:
            x-session-token:
              schema:
                type: string
              description: The JWT associated with this authenticated session
        "400":
          description: User signup request malformed or denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorModel"

  # Read or change a user
...
  /billing_portal:
    get:
      summary: Redirects to your Stripe customer billing portal
      operationId: viewBillingPortal
      responses:
        "301":
          description: "Portal redirect success"
        "400":
          description: "Error redirecting to billing portal"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorModel"

  # Sign up as tutor
  /tutor:
    post:
      summary: Create Tutor Profile for User
      operationId: signUpAsTutor
      requestBody:
        required: false
        content:
          application/json: {}
      security:
        - BearerAuth: []
      responses:
        "201":
          description: Tutor signed up
        "400":
          description: Create tutor request malformed or denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
    get:
      summary: Search All Tutors
      operationId: getTutors
      parameters:
        - name: pageSize
          in: query
          required: true
          description: The ID of the tutor to get
          schema:
            type: integer
        - name: pageIndex
          in: query
          required: true
          description: The ID of the tutor to get
          schema:
            type: integer
        - name: minRating
          in: query
          required: false
          description: The minimum rating of tutor to get.
          schema:
            type: number
        - name: skillsInclude
          in: query
          required: false
          description: The skills a tutor should have.
          schema:
            type: array
            items:
              type: string
              format: uuid

      responses:
        "200":
          description: Found tutors.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    info:
                      $ref: "#/components/schemas/Tutor"
                    ratingScores:
                      $ref: "#/components/schemas/RatingScores"
                    ratingCount:
                      type: integer
        "400":
          description: Get tutor request malformed or denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
        "403":
          description: Not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
        "404":
          description: Tutors not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorModel"

  # Read or change tutor information
  /tutor/{tutorId}:
    get:
      summary: Get a tutor by ID
      operationId: getTutorById
      parameters:
        - name: tutorId
          in: path
          required: true
          description: The ID of the tutor to get
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Tutor found
          content:
            application/json:
              schema:
                type: object
                properties:
                  info:
                    $ref: "#/components/schemas/Tutor"
                  ratingScores:
                    $ref: "#/components/schemas/RatingScores"
                  ratingCount:
                    type: integer
        "400":
          description: Get tutor request malformed or denied.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
        "404":
          description: Tutor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorModel"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer

    GitHubOAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://github.com/login/oauth/authorize
          tokenUrl: https://github.com/login/oauth/access_token
          scopes:
            read:user: "Grants read access to the user's profile data on GitHub"
            repo: "Grants read and write access to public and private repositories"

    GoogleOAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://accounts.google.com/o/oauth2/auth
          tokenUrl: https://oauth2.googleapis.com/token
          scopes:
            openid: "Grants access to user's basic profile information"
            email: "Grants access to the user's email address"
            profile: "Grants access to the user's profile information"
  schemas:
    ErrorModel:
      type: object
      required:
        - message
        - code
      example:
        message: Resource not found.
        code: 400
      properties:
        message:
          type: string
        code:
          type: integer
          minimum: 100
          maximum: 600
    User:
      type: object
      example:
        userId: 78fcb19e-5ae8-4c5f-a468-b08135a16235
        email: email@example.com
        firstName: Jacob
        lastName: Hochstetler
      description: Base User object containing shared details needed for all users.
      required:
        - userId
        - firstName
        - lastName
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
          format: email
        signedUpAt:
          type: string
          format: date-time
        firstName:
          type: string
        lastName:
          type: string
        accountLocked:
          type: boolean
        passwordHash:
          type: string
    UserLogin:
      type: object
      example:
        email: email@example.com
        password: N3v3rG0nn4G1veY0uUp!
      description: Payload for user logins
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        rememberLogin:
          type: boolean
    UserSignup:
      type: object
      example:
        email: email@example.com
        password: N3v3rG0nn4G1veY0uUp!
        firstName: Jacob
        lastName: Hochstetler
      description: Payload for user signups
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        first_name:
          type: string
        last_name:
          type: string
    Tutor:
      description: Extends user to include student specific information.
      allOf:
        - $ref: "#/components/schemas/User"
        - type: object
          example:
            userId: 78fcb19e-5ae8-4c5f-a468-b08135a16235
          properties:
            totalHours: # total hours worked
              type: integer
            skills:
              type: array
              items:
                type: string
    RatingScores:
      type: object
      required:
        - overall
        - professionalism
        - knowledge
        - communication
        - punctuality
      properties: # Subject to change?
        professionalism:
          type: integer
          minimum: 1
          maximum: 5
        knowledge:
          type: integer
          minimum: 1
          maximum: 5
        communication:
          type: integer
          minimum: 1
          maximum: 5
        punctuality:
          type: integer
          minimum: 1
          maximum: 5
        overall:
          type: integer
          minimum: 1
          maximum: 5
...